Refactor contextual queries
Break the  file into logically contained
smaller files. And add unit tests for .
